requirements.txt
flask==3.0.0
gunicorn==21.2.0
groq==0.4.2
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.0
